+++ web/e2e/project/utils.ts
@@ -2,21 +2,6 @@ import { Page } from "@playwright/test";
 
 import { expect } from "@reearth-cms/e2e/utils";
 
-export async function createWorkspace(page: Page) {
-  await page.getByRole("button", { name: "Create a Workspace" }).click();
-  await page.getByLabel("Workspace name").click();
-  await page.getByLabel("Workspace name").fill("e2e workspace name");
-  await page.getByRole("button", { name: "OK" }).click();
-  await expect(page.getByRole("alert").last()).toContainText("Successfully created workspace!");
-}
-
-export async function deleteWorkspace(page: Page) {
-  await page.getByText("Workspace", { exact: true }).click();
-  await page.getByRole("button", { name: "Remove Workspace" }).click();
-  await page.getByRole("button", { name: "OK" }).click();
-  await expect(page.getByRole("alert").last()).toContainText("Successfully deleted workspace!");
-}
-
 export async function createProject(page: Page) {
   await page.getByRole("button", { name: "plus New Project" }).first().click();
   await page.getByRole("dialog").locator("#name").click();
